+++ maps-utils/src/main/java/com/google/maps/android/SphericalUtil.kt
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2013 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.maps.android
+
+import com.google.maps.android.MathUtil.EARTH_RADIUS
+import com.google.maps.android.MathUtil.arcHav
+import com.google.maps.android.MathUtil.havDistance
+import com.google.maps.android.model.LatLng
+
+object SphericalUtil {
+    /**
+     * Returns distance on the unit sphere; the arguments are in radians.
+     */
+    private fun distanceRadians(lat1: Double, lng1: Double, lat2: Double, lng2: Double): Double {
+        return arcHav(havDistance(lat1, lat2, lng1 - lng2))
+    }
+
+    /**
+     * Returns the angle between two LatLngs, in radians. This is the same as the distance
+     * on the unit sphere.
+     */
+    fun computeAngleBetween(from: LatLng, to: LatLng): Double {
+        return distanceRadians(
+            Math.toRadians(from.latitude), Math.toRadians(from.longitude),
+            Math.toRadians(to.latitude), Math.toRadians(to.longitude)
+        )
+    }
+
+    /**
+     * Returns the distance between two LatLngs, in meters.
+     */
+    fun computeDistanceBetween(from: LatLng, to: LatLng): Double {
+        return computeAngleBetween(from, to) * EARTH_RADIUS
+    }
+}